Rename PhantomCoreOrchestrator file name and update tests
